pyproject.toml
[tool.poetry]
name ="movie_app"
version = "0.1.0"
description = "fastAPI"
authors = [
    "yas <you@example.com>"
]
readme = "README.md"


packages = [{include = "movie_app", from = "src"}]


[tool.poetry.dependencies]
fastapi = { extras = ["standard"], version = ">=0.116.1,<0.117.0" }
sqlalchemy = ">=2.0.42,<3.0.0"
alembic = ">=1.16.4,<2.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
pydantic = ">=2.6,<3.0"
pydantic-settings = ">=2.10.1,<3.0.0"
python = ">=3.11"
psycopg2-binary = "^2.9.10"
redis = "^6.2.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}


[tool.poetry.group.dev.dependencies]
sqlalchemy2-stubs = "^0.0.2a38"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
